Move extract_package into the cgo wrapper
This is needed in order to stop using bash, which is needed to enable the new go
environment handling and also to fix windows builds.
